+++ tests/test_files_outputs_parity.py
@@ -21,14 +21,16 @@ LEGACY_FILEENTRY_KEYS = {
     "name", "path", "type", "size", "extension", "convertible", "has_chd",
     "has_rvz", "dolphin_ready", "dolphin_path", "chd_ready",
     "dolphin_convertible", "z3ds_convertible", "has_z3ds", "z3ds_ready",
-    "z3ds_path", "archive_items", "archive_has_chd", "archive_truncated",
+    "z3ds_path", "nsz_convertible", "has_nsz", "nsz_ready", "nsz_path",
+    "archive_items", "archive_has_chd", "archive_truncated",
     "media_type",
 }
 LEGACY_SEARCH_KEYS = {
     "name", "path", "size", "extension", "chd_path", "has_chd", "has_rvz",
     "dolphin_ready", "dolphin_path", "chd_ready", "convertible",
     "dolphin_convertible", "z3ds_convertible", "has_z3ds", "z3ds_ready",
-    "z3ds_path", "in_archive",
+    "z3ds_path", "nsz_convertible", "has_nsz", "nsz_ready", "nsz_path",
+    "in_archive",
 }
 NEW_KEYS = {"convertible_by", "outputs"}